refactor: app.jsをES modulesで6ファイルに分割
3538行の巨大なapp.jsを責務ごとにモジュール分割。
ビルドツール不要、ブラウザネイティブES modulesのまま動作。

- analysis/stats.js: 分析関数16種 + ヘルパー
- analysis/aggregate.js: 全機体集約関数7種
- lib/db.js: IndexedDBキャッシュ
- lib/format.js: 色分け・フォーマット・SNS共有
- components/ui.js: 汎用UIコンポーネント (Tips, Table等)
- components/charts.js: Chart.jsグラフ6種 + レポートセクション9種
