+++ static/app.js
@@ -1,4 +1,38 @@
 import { html, render, useState, useMemo, useCallback, useEffect, useRef } from './htm-preact-standalone.js';
+import {
+  PERIOD_DAYS, filterByPlayDays,
+  computeTimeOfDay, computeDayOfWeek, computeDailyTrend, computeSeason,
+  computeBasicStats, computeWinLossPattern,
+  computeEnemyMatchup, computePartner, computeCostPair, computeMsPair,
+  computeDmgContribution, computeDeathsImpact,
+  computeBurstCount, computeFallOrder, computeBurstHoldDeath,
+  computeFixedPartners,
+} from './analysis/stats.js';
+import {
+  aggregateDeathsImpact, aggregateFallOrder, aggregateDmgContribution,
+  aggregateBurstCount, aggregateBurstHoldDeath,
+  aggregateEnemyMatchup, aggregatePartner,
+} from './analysis/aggregate.js';
+import {
+  loadMatchesFromDB, saveMatchesToDB, fetchAndCacheMatches,
+} from './lib/db.js';
+import {
+  esc, pct, num, colorPct, colorDE,
+  colorDmgGiven, colorDmgTaken, colorKills, colorDeaths, colorKD,
+  colorExDmg, colorBursts, cellDisplay,
+  buildShareText, SVG_X, SVG_BSKY, SVG_LINE, SVG_COPY, SVG_CHECK,
+} from './lib/format.js';
+import {
+  Tips, SortableTable, Table, SubSection,
+} from './components/ui.js';
+import {
+  useInView,
+  EnemyMatchupSection, PartnerSection, MsPairSubSection, CostPairSubSection,
+  DmgContributionSubSection, DeathsImpactSubSection,
+  TimeOfDayChart, DayOfWeekChart, DailyTrendChart, SeasonChart,
+  WinRateBarChart, DmgContributionChart,
+  FallOrderContent, BurstHoldDeathContent, BurstCountContent,
+} from './components/charts.js';
 
 // --- Constants ---
 var STATUS_MESSAGES = {
@@ -11,1235 +45,6 @@ var STATUS_MESSAGES = {
 };
 
 var PERIOD_KEYS = ['all', '90d', '60d', '30d', '14d', '7d', '3d', '1d'];
-
-// --- IndexedDB Match Cache ---
-var MATCH_DB_NAME = 'catalyzer';
-var MATCH_DB_VERSION = 1;
-var MATCH_STORE = 'matches';
-
-function openMatchDB() {
-  return new Promise(function (resolve, reject) {
-    var req = indexedDB.open(MATCH_DB_NAME, MATCH_DB_VERSION);
-    req.onupgradeneeded = function (e) {
-      var db = e.target.result;
-      if (!db.objectStoreNames.contains(MATCH_STORE)) {
-        var store = db.createObjectStore(MATCH_STORE, { keyPath: 'id' });
-        store.createIndex('userKey', 'user_key', { unique: false });
-        store.createIndex('date', 'date', { unique: false });
-        store.createIndex('ms', 'ms', { unique: false });
-        store.createIndex('userKey_date', ['user_key', 'date'], { unique: false });
-      }
-    };
-    req.onsuccess = function (e) { resolve(e.target.result); };
-    req.onerror = function (e) { reject(e.target.error); };
-  });
-}
-
-function saveMatchesToDB(userKey, matches) {
-  return openMatchDB().then(function (db) {
-    return new Promise(function (resolve, reject) {
-      var tx = db.transaction(MATCH_STORE, 'readwrite');
-      var store = tx.objectStore(MATCH_STORE);
-      for (var i = 0; i < matches.length; i++) {
-        var m = Object.assign({}, matches[i], {
-          id: userKey + '_' + matches[i].date,
-          user_key: userKey
-        });
-        store.put(m);
-      }
-      tx.oncomplete = function () { resolve(); };
-      tx.onerror = function (e) { reject(e.target.error); };
-    });
-  });
-}
-
-function loadMatchesFromDB(userKey) {
-  return openMatchDB().then(function (db) {
-    return new Promise(function (resolve, reject) {
-      var tx = db.transaction(MATCH_STORE, 'readonly');
-      var store = tx.objectStore(MATCH_STORE);
-      var index = store.index('userKey');
-      var req = index.getAll(userKey);
-      req.onsuccess = function (e) { resolve(e.target.result || []); };
-      req.onerror = function (e) { reject(e.target.error); };
-    });
-  });
-}
-
-function getLatestMatchDate(userKey) {
-  return openMatchDB().then(function (db) {
-    return new Promise(function (resolve, reject) {
-      var tx = db.transaction(MATCH_STORE, 'readonly');
-      var store = tx.objectStore(MATCH_STORE);
-      var index = store.index('userKey_date');
-      var range = IDBKeyRange.bound([userKey, ''], [userKey, '￿']);
-      var req = index.openCursor(range, 'prev');
-      req.onsuccess = function (e) {
-        var cursor = e.target.result;
-        resolve(cursor ? cursor.value.date : null);
-      };
-      req.onerror = function (e) { reject(e.target.error); };
-    });
-  });
-}
-
-function fetchAndCacheMatches(userKey) {
-  getLatestMatchDate(userKey).then(function (latestDate) {
-    var url = '/matches?user_key=' + encodeURIComponent(userKey);
-    if (latestDate) {
-      url += '&after=' + encodeURIComponent(latestDate);
-    }
-    return fetch(url);
-  }).then(function (res) {
-    if (!res.ok) throw new Error('fetch failed');
-    return res.json();
-  }).then(function (data) {
-    if (data.matches && data.matches.length > 0) {
-      return saveMatchesToDB(userKey, data.matches);
-    }
-  }).catch(function () {});
-}
-
-// --- Frontend Aggregation Helpers ---
-var PERIOD_DAYS = { '90d': 90, '60d': 60, '30d': 30, '14d': 14, '7d': 7, '3d': 3, '1d': 1 };
-
-function jsWinRate(matches) {
-  if (!matches.length) return 0;
-  var w = 0;
-  for (var i = 0; i < matches.length; i++) { if (matches[i].win) w++; }
-  return w / matches.length * 100;
-}
-
-function jsDmgEfficiency(matches) {
-  if (!matches.length) return 0;
-  var g = 0, t = 0;
-  for (var i = 0; i < matches.length; i++) { g += matches[i].dmg_given; t += matches[i].dmg_taken; }
-  return t > 0 ? g / t : 0;
-}
-
-function round1(n) { return Math.round(n * 10) / 10; }
-function round3(n) { return Math.round(n * 1000) / 1000; }
-
-function filterByPlayDays(matches, days) {
-  if (!matches.length) return [];
-  var dateSet = {};
-  for (var i = 0; i < matches.length; i++) {
-    dateSet[matches[i].date.substring(0, 10)] = true;
-  }
-  var playDates = Object.keys(dateSet).sort().reverse();
-  var targetDates = {};
-  for (var i = 0; i < Math.min(days, playDates.length); i++) {
-    targetDates[playDates[i]] = true;
-  }
-  return matches.filter(function (m) { return targetDates[m.date.substring(0, 10)]; });
-}
-
-function computeTimeOfDay(matches) {
-  var hourly = {};
-  for (var i = 0; i < matches.length; i++) {
-    var hour = parseInt(matches[i].date.substring(11, 13), 10);
-    if (!hourly[hour]) hourly[hour] = [];
-    hourly[hour].push(matches[i]);
-  }
-  var hours = [];
-  for (var h = 0; h < 24; h++) {
-    if (!hourly[h]) continue;
-    var ms = hourly[h];
-    var wr = jsWinRate(ms);
-    hours.push({ hour: h, matches: ms.length, win_rate: round1(wr), dmg_efficiency: round3(jsDmgEfficiency(ms)), mark: wr >= 70 ? 'good' : wr <= 40 ? 'bad' : '' });
-  }
-  var tips = [];
-  var good = [], bad = [];
-  for (var h in hourly) {
-    if (hourly[h].length >= 5) {
-      var wr = jsWinRate(hourly[h]);
-      if (wr >= 70) good.push(Number(h));
-      if (wr <= 40) bad.push(Number(h));
-    }
-  }
-  good.sort(function (a, b) { return a - b; });
-  bad.sort(function (a, b) { return a - b; });
-  if (good.length) tips.push('好調 → **' + good.map(function (h) { return h + '時台'; }).join('、') + '**');
-  if (bad.length) tips.push('不調 → **' + bad.map(function (h) { return h + '時台'; }).join('、') + '**（強豪が多い or 疲労の影響）');
-  return { hours: hours, tips: tips };
-}
-
-function computeDayOfWeek(matches) {
-  var DOW_NAMES = ['月', '火', '水', '木', '金', '土', '日'];
-  var daily = {};
-  var weekdayData = [], weekendData = [];
-  for (var i = 0; i < matches.length; i++) {
-    var m = matches[i];
-    var d = new Date(m.date.replace(' ', 'T'));
-    var dow = (d.getDay() + 6) % 7;
-    if (!daily[dow]) daily[dow] = [];
-    daily[dow].push(m);
-    if (dow < 5) weekdayData.push(m); else weekendData.push(m);
-  }
-  var days = [];
-  for (var dow = 0; dow < 7; dow++) {
-    if (daily[dow]) {
-      days.push({ dow: dow, name: DOW_NAMES[dow], matches: daily[dow].length, win_rate: round1(jsWinRate(daily[dow])), dmg_efficiency: round3(jsDmgEfficiency(daily[dow])) });
-    }
-  }
-  var wdWr = weekdayData.length ? jsWinRate(weekdayData) : 0;
-  var weWr = weekendData.length ? jsWinRate(weekendData) : 0;
-  var diff = Math.abs(wdWr - weWr);
-  var tips = [];
-  if (diff >= 10) {
-    var better = wdWr > weWr ? '平日' : '土日';
-    var worse = wdWr > weWr ? '土日' : '平日';
-    tips.push('**' + better + '**が' + worse + 'より勝率 **' + Math.round(diff) + '%** 高い');
-  }
-  return {
-    weekday: weekdayData.length ? { matches: weekdayData.length, win_rate: round1(wdWr), dmg_efficiency: round3(jsDmgEfficiency(weekdayData)) } : { matches: 0, win_rate: 0, dmg_efficiency: 0 },
-    weekend: weekendData.length ? { matches: weekendData.length, win_rate: round1(weWr), dmg_efficiency: round3(jsDmgEfficiency(weekendData)) } : { matches: 0, win_rate: 0, dmg_efficiency: 0 },
-    days: days,
-    tips: tips,
-  };
-}
-
-function computeDailyTrend(matches) {
-  var DOW_NAMES = ['月', '火', '水', '木', '金', '土', '日'];
-  var daily = {};
-  var dateOrder = {};
-  for (var i = 0; i < matches.length; i++) {
-    var m = matches[i];
-    var dateKey = m.date.substring(5, 10).replace('-', '/');
-    if (!daily[dateKey]) { daily[dateKey] = []; dateOrder[dateKey] = m.date.substring(0, 10); }
-    daily[dateKey].push(m);
-  }
-  var sortedKeys = Object.keys(daily).sort(function (a, b) {
-    return dateOrder[a] < dateOrder[b] ? -1 : dateOrder[a] > dateOrder[b] ? 1 : 0;
-  });
-  var results = sortedKeys.map(function (dateStr) {
-    var ms = daily[dateStr];
-    var wr = jsWinRate(ms);
-    var d = new Date(ms[0].date.replace(' ', 'T'));
-    var dow = (d.getDay() + 6) % 7;
-    return { date: dateStr, dow_name: DOW_NAMES[dow], matches: ms.length, win_rate: round1(wr), dmg_efficiency: round3(jsDmgEfficiency(ms)), mark: wr >= 70 ? 'good' : wr <= 45 ? 'bad' : '' };
-  });
-  var tips = [];
-  var badDays = sortedKeys.filter(function (ds) { return jsWinRate(daily[ds]) <= 40 && daily[ds].length >= 5; });
-  if (badDays.length) tips.push('不調日（勝率40%以下）→ **' + badDays.join(', ') + '**。早めの切り上げが有効');
-  return { days: results, tips: tips };
-}
-
-// --- Phase 2 Frontend Aggregation ---
-var COST_LABEL = {3000: '3000コスト', 2500: '2500コスト', 2000: '2000コスト', 1500: '1500コスト'};
-var COST_FATAL_DEATHS = {3000: 2, 2500: 3, 2000: 3, 1500: 4};
-
-function jsAvg(arr) { return arr.length ? arr.reduce(function (a, b) { return a + b; }, 0) / arr.length : 0; }
-function jsWinsLosses(ms) { var w = ms.filter(function (m) { return m.win; }).length; return [w, ms.length - w]; }
-function jsKdRatio(ms) { var k = 0, d = 0; ms.forEach(function (m) { k += m.kills; d += m.deaths; }); return d > 0 ? k / d : 0; }
-function jsAvgBursts(ms) {
-  var valid = ms.filter(function (m) { return m.actions && m.actions.length > 0; });
-  if (!valid.length) return null;
-  return jsAvg(valid.map(function (m) { return m.bursts; }));
-}
-function round2(n) { return Math.round(n * 100) / 100; }
-
-function computeBasicStats(matches) {
-  var n = matches.length;
-  var wl = jsWinsLosses(matches);
-  var w = wl[0], l = wl[1];
-  var wr = n > 0 ? w / n * 100 : 0;
-  var kd = jsKdRatio(matches);
-  var eff = jsDmgEfficiency(matches);
-  var ab = jsAvgBursts(matches);
-
-  var tips = [];
-  if (eff < 1.0) {
-    tips.push('与被ダメ比 **' + round3(eff) + '** → 被ダメ超過。被弾を減らす立ち回りを意識');
-  } else if (eff >= 1.2) {
-    tips.push('与被ダメ比 **' + round3(eff) + '** → 優秀。この調子を維持');
-  }
-  if (kd < 1.0) {
-    tips.push('K/D比 **' + round2(kd) + '** → 1.0未満。撃墜↑ or 被撃墜↓を意識');
-  }
-
-  return {
-    matches: n,
-    wins: w,
-    losses: l,
-    win_rate: round1(wr),
-    avg_dmg_given: Math.round(jsAvg(matches.map(function (m) { return m.dmg_given; }))),
-    avg_dmg_taken: Math.round(jsAvg(matches.map(function (m) { return m.dmg_taken; }))),
-    dmg_efficiency: round3(eff),
-    avg_kills: round2(jsAvg(matches.map(function (m) { return m.kills; }))),
-    avg_deaths: round2(jsAvg(matches.map(function (m) { return m.deaths; }))),
-    kd_ratio: round2(kd),
-    avg_ex_dmg: Math.round(jsAvg(matches.map(function (m) { return m.ex_dmg; }))),
-    avg_bursts: ab !== null ? round2(ab) : null,
-    tips: tips,
-  };
-}
-
-function computeWinLossPattern(matches) {
-  var wins = matches.filter(function (m) { return m.win; });
-  var losses = matches.filter(function (m) { return !m.win; });
-
-  function kdOf(rows) {
-    var tk = 0, td = 0;
-    rows.forEach(function (d) { tk += d.kills; td += d.deaths; });
-    return td > 0 ? tk / td : 0;
-  }
-  function avgOf(key, rows) { return rows.length ? jsAvg(rows.map(function (d) { return d[key]; })) : 0; }
-
-  var metrics = [];
-  function addMetric(label, w, l, nd) {
-    var factor = Math.pow(10, nd);
-    metrics.push({
-      label: label,
-      win_avg: w !== null ? Math.round(w * factor) / factor : null,
-      loss_avg: l !== null ? Math.round(l * factor) / factor : null,
-      diff: (w !== null && l !== null) ? Math.round((w - l) * factor) / factor : null,
-    });
-  }
-
-  addMetric('平均与ダメージ', avgOf('dmg_given', wins), avgOf('dmg_given', losses), 1);
-  addMetric('平均被ダメージ', avgOf('dmg_taken', wins), avgOf('dmg_taken', losses), 1);
-  addMetric('与被ダメ比', wins.length ? jsDmgEfficiency(wins) : 0, losses.length ? jsDmgEfficiency(losses) : 0, 3);
-  addMetric('平均撃墜', avgOf('kills', wins), avgOf('kills', losses), 2);
-  addMetric('平均被撃墜', avgOf('deaths', wins), avgOf('deaths', losses), 2);
-  addMetric('K/D比', wins.length ? kdOf(wins) : 0, losses.length ? kdOf(losses) : 0, 2);
-  addMetric('平均EXダメージ', avgOf('ex_dmg', wins), avgOf('ex_dmg', losses), 1);
-  addMetric('平均覚醒回数', jsAvgBursts(wins), jsAvgBursts(losses), 2);
-
-  var tips = [];
-  var lDeaths = losses.length ? jsAvg(losses.map(function (d) { return d.deaths; })) : 0;
-  var lTaken = losses.length ? jsAvg(losses.map(function (d) { return d.dmg_taken; })) : 0;
-  if (lDeaths >= 1.5) {
-    tips.push('敗北時の平均被撃墜 **' + round1(lDeaths) + '回** → 耐久管理を意識');
-  }
-  if (lTaken >= 1100) {
-    tips.push('敗北時の平均被ダメ **' + Math.round(lTaken) + '** → 無駄な被弾を減らすのが改善の鍵');
-  }
-
-  // コスト帯別
-  var costGroups = {};
-  matches.forEach(function (d) {
-    var cost = d.ms_cost || 0;
-    if (COST_LABEL[cost]) {
-      if (!costGroups[cost]) costGroups[cost] = [];
-      costGroups[cost].push(d);
-    }
-  });
-
-  var costPatterns = [];
-  var costKeys = Object.keys(costGroups).map(Number).sort(function (a, b) { return b - a; });
-  costKeys.forEach(function (cost) {
-    var data = costGroups[cost];
-    if (data.length < 3) return;
-    var cWins = data.filter(function (d) { return d.win; });
-    var cLosses = data.filter(function (d) { return !d.win; });
-    if (!cWins.length || !cLosses.length) return;
-
-    var fatal = COST_FATAL_DEATHS[cost];
-    var fatalLosses = cLosses.filter(function (d) { return d.deaths >= fatal; });
-
-    var costMetrics = [];
-    [['与ダメージ', 'dmg_given'], ['被ダメージ', 'dmg_taken'], ['撃墜', 'kills'], ['被撃墜', 'deaths']].forEach(function (pair) {
-      var mLabel = pair[0], key = pair[1];
-      var wV = jsAvg(cWins.map(function (d) { return d[key]; }));
-      var lV = jsAvg(cLosses.map(function (d) { return d[key]; }));
-      costMetrics.push({ label: mLabel, win_avg: round1(wV), loss_avg: round1(lV), diff: round1(wV - lV) });
-    });
-    var cWEff = jsDmgEfficiency(cWins);
-    var cLEff = jsDmgEfficiency(cLosses);
-    costMetrics.push({ label: '与被ダメ比', win_avg: round3(cWEff), loss_avg: round3(cLEff), diff: round3(cWEff - cLEff) });
-
-    costPatterns.push({
-      cost: cost,
-      cost_label: COST_LABEL[cost],
-      matches: data.length,
-      win_rate: round1(jsWinRate(data)),
-      metrics: costMetrics,
-      fatal_deaths: fatal,
-      fatal_loss_count: fatalLosses.length,
-      fatal_loss_total: cLosses.length,
-      fatal_loss_rate: cLosses.length ? Math.round(fatalLosses.length / cLosses.length * 100) : 0,
-    });
-  });
-
-  return { metrics: metrics, tips: tips, cost_patterns: costPatterns };
-}
-
-function computeEnemyMatchup(matches, minMatches) {
-  if (minMatches === undefined) minMatches = 3;
-  var enemyStats = {};
-  matches.forEach(function (d) {
-    [d.opponent1_ms, d.opponent2_ms].forEach(function (ems) {
-      if (!ems) return;
-      if (!enemyStats[ems]) enemyStats[ems] = [];
-      enemyStats[ems].push(d);
-    });
-  });
-
-  var results = [];
-  Object.keys(enemyStats).forEach(function (ms) {
-    var ms_matches = enemyStats[ms];
-    if (ms_matches.length >= minMatches) {
-      var wr = jsWinRate(ms_matches);
-      var eff = jsDmgEfficiency(ms_matches);
-      var avgGiven = jsAvg(ms_matches.map(function (d) { return d.dmg_given; }));
-      var avgTaken = jsAvg(ms_matches.map(function (d) { return d.dmg_taken; }));
-      results.push({
-        ms: ms,
-        matches: ms_matches.length,
-        win_rate: round1(wr),
-        dmg_efficiency: round3(eff),
-        avg_dmg_given: Math.round(avgGiven),
-        avg_dmg_taken: Math.round(avgTaken),
-      });
-    }
-  });
-
-  var strong = results.filter(function (r) { return r.win_rate >= 60; }).sort(function (a, b) { return b.matches - a.matches; });
-  var weak = results.filter(function (r) { return r.win_rate <= 40; }).sort(function (a, b) { return b.matches - a.matches; });
-  var even = results.filter(function (r) { return r.win_rate > 40 && r.win_rate < 60; }).sort(function (a, b) { return b.matches - a.matches; });
-
-  var tips = [];
-  if (weak.length) {
-    var highDmgTaken = weak.filter(function (r) { return r.avg_dmg_taken >= 1200; });
-    if (highDmgTaken.length) {
-      tips.push({ text: '被ダメが多い相手 → 距離管理を見直し', details: highDmgTaken.slice(0, 3).map(function (r) { return '**' + r.ms + '** 被ダメ ' + r.avg_dmg_taken; }) });
-    }
-    var lowDmgGiven = weak.filter(function (r) { return r.avg_dmg_given <= 900; });
-    if (lowDmgGiven.length) {
-      tips.push({ text: '与ダメが低い相手 → 手数や当て方を工夫', details: lowDmgGiven.slice(0, 3).map(function (r) { return '**' + r.ms + '** 与ダメ ' + r.avg_dmg_given; }) });
-    }
-  }
-
-  return { strong: strong, weak: weak, even: even, tips: tips };
-}
-
-function computePartner(matches, minMatches) {
-  if (minMatches === undefined) minMatches = 3;
-  var partnerStats = {};
-  matches.forEach(function (d) {
-    var pms = d.partner_ms;
-    if (!pms) return;
-    if (!partnerStats[pms]) partnerStats[pms] = [];
-    partnerStats[pms].push(d);
-  });
-
-  var results = [];
-  Object.keys(partnerStats).forEach(function (ms) {
-    var ms_matches = partnerStats[ms];
-    if (ms_matches.length >= minMatches) {
-      results.push({
-        ms: ms,
-        matches: ms_matches.length,
-        win_rate: round1(jsWinRate(ms_matches)),
-        dmg_efficiency: round3(jsDmgEfficiency(ms_matches)),
-      });
-    }
-  });
-  results.sort(function (a, b) { return b.matches - a.matches; });
-  return results;
-}
-
-function computeCostPair(matches, minMatches) {
-  if (minMatches === undefined) minMatches = 3;
-  var pairs = {};
-  matches.forEach(function (d) {
-    var msName = d.ms || '(不明)';
-    var partnerCost = d.partner_cost || 0;
-    if (msName && partnerCost) {
-      var key = msName + ' + ' + partnerCost;
-      if (!pairs[key]) pairs[key] = [];
-      pairs[key].push(d);
-    }
-  });
-
-  var results = [];
-  Object.keys(pairs).forEach(function (pair) {
-    var ms_matches = pairs[pair];
-    if (ms_matches.length >= minMatches) {
-      results.push({
-        pair: pair,
-        matches: ms_matches.length,
-        win_rate: round1(jsWinRate(ms_matches)),
-        dmg_efficiency: round3(jsDmgEfficiency(ms_matches)),
-      });
-    }
-  });
-  results.sort(function (a, b) { return b.matches - a.matches; });
-  return results;
-}
-
-function computeMsPair(matches, minMatches, topN) {
-  if (minMatches === undefined) minMatches = 3;
-  if (topN === undefined) topN = 10;
-  var pairs = {};
-  matches.forEach(function (d) {
-    var key = d.ms + ' + ' + d.partner_ms;
-    if (!pairs[key]) pairs[key] = [];
-    pairs[key].push(d);
-  });
-
-  var results = [];
-  Object.keys(pairs).forEach(function (pair) {
-    var ms_matches = pairs[pair];
-    if (ms_matches.length >= minMatches) {
-      var wl = jsWinsLosses(ms_matches);
-      results.push({
-        pair: pair,
-        matches: ms_matches.length,
-        wins: wl[0],
-        losses: wl[1],
-        win_rate: round1(jsWinRate(ms_matches)),
-        dmg_efficiency: round3(jsDmgEfficiency(ms_matches)),
-      });
-    }
-  });
-
-  var byWr = results.slice().sort(function (a, b) { return a.win_rate !== b.win_rate ? b.win_rate - a.win_rate : b.matches - a.matches; }).slice(0, topN);
-  var byCount = results.slice().sort(function (a, b) { return a.matches !== b.matches ? b.matches - a.matches : b.win_rate - a.win_rate; }).slice(0, topN);
-
-  return { by_win_rate: byWr, by_matches: byCount };
-}
-
-function computeDmgContribution(matches, minMatches) {
-  if (minMatches === undefined) minMatches = 3;
-  var contribs = [];
-  var winContribs = [];
-  var loseContribs = [];
-  matches.forEach(function (d) {
-    var teamTotal = d.dmg_given + d.partner_dmg_given;
-    if (teamTotal > 0) {
-      var c = d.dmg_given / teamTotal * 100;
-      contribs.push(c);
-      if (d.win) { winContribs.push(c); } else { loseContribs.push(c); }
-    }
-  });
-  var avgContrib = contribs.length ? jsAvg(contribs) : 0;
-  var avgWin = winContribs.length ? jsAvg(winContribs) : 0;
-  var avgLose = loseContribs.length ? jsAvg(loseContribs) : 0;
-
-  // コスト帯別
-  var costGroups = {};
-  matches.forEach(function (d) {
-    var cost = d.ms_cost || 0;
-    if (COST_LABEL[cost]) {
-      if (!costGroups[cost]) costGroups[cost] = [];
-      costGroups[cost].push(d);
-    }
-  });
-
-  var costData = [];
-  var costKeys = Object.keys(costGroups).map(Number);
-  var hasMultiple = costKeys.length > 1 || costKeys.some(function (k) { return costGroups[k].length >= minMatches; });
-  if (hasMultiple) {
-    costKeys.sort(function (a, b) { return b - a; }).forEach(function (cost) {
-      var data = costGroups[cost];
-      if (data.length < minMatches) return;
-      var cAll = [], cWin = [], cLose = [];
-      data.forEach(function (d) {
-        var teamTotal = d.dmg_given + d.partner_dmg_given;
-        if (teamTotal > 0) {
-          var c = d.dmg_given / teamTotal * 100;
-          cAll.push(c);
-          if (d.win) { cWin.push(c); } else { cLose.push(c); }
-        }
-      });
-      costData.push({
-        cost: cost,
-        cost_label: COST_LABEL[cost],
-        matches: data.length,
-        avg_contribution: cAll.length ? round1(jsAvg(cAll)) : 0,
-        avg_win_contribution: cWin.length ? round1(jsAvg(cWin)) : 0,
-        avg_lose_contribution: cLose.length ? round1(jsAvg(cLose)) : 0,
-      });
-    });
-  }
-
-  return {
-    avg_contribution: round1(avgContrib),
-    avg_win_contribution: round1(avgWin),
-    avg_lose_contribution: round1(avgLose),
-    by_cost: costData,
-  };
-}
-
-function computeDeathsImpact(matches) {
-  var costGroups = {};
-  matches.forEach(function (d) {
-    var cost = d.ms_cost || 0;
-    if (COST_FATAL_DEATHS[cost]) {
-      if (!costGroups[cost]) costGroups[cost] = [];
-      costGroups[cost].push(d);
-    }
-  });
-
-  var results = [];
-  var costKeys = Object.keys(costGroups).map(Number).sort(function (a, b) { return b - a; });
-  costKeys.forEach(function (cost) {
-    var data = costGroups[cost];
-    var fatal = COST_FATAL_DEATHS[cost];
-    var maxBucket = fatal + 1;
-
-    var byDeaths = {};
-    data.forEach(function (d) {
-      var deaths = d.deaths;
-      var key = deaths >= maxBucket ? (maxBucket + '+') : String(deaths);
-      if (!byDeaths[key]) byDeaths[key] = [];
-      byDeaths[key].push(d);
-    });
-
-    var buckets = [];
-    for (var i = 0; i < fatal; i++) {
-      var key = String(i);
-      if (byDeaths[key]) {
-        var bMatches = byDeaths[key];
-        var wr = jsWinRate(bMatches);
-        var eff = jsDmgEfficiency(bMatches);
-        buckets.push({
-          deaths: i,
-          label: i + '落ち',
-          matches: bMatches.length,
-          win_rate: round1(wr),
-          dmg_efficiency: round3(eff),
-          status: i === fatal - 1 ? 'danger' : 'safe',
-        });
-      }
-    }
-
-    var fatalCount = 0;
-    for (var j = fatal; j < maxBucket; j++) {
-      fatalCount += (byDeaths[String(j)] || []).length;
-    }
-    fatalCount += (byDeaths[maxBucket + '+'] || []).length;
-
-    var safeData = [];
-    for (var k = 0; k < fatal; k++) {
-      safeData = safeData.concat(byDeaths[String(k)] || []);
-    }
-    var safeWr = safeData.length ? jsWinRate(safeData) : 0;
-
-    var tips = [];
-    if (fatalCount > 0 && data.length > 0) {
-      tips.push(fatal + '落ち以上 **' + fatalCount + '/' + data.length + '戦**（' + Math.round(fatalCount / data.length * 100) + '%）');
-    }
-    if (safeData.length) {
-      tips.push((fatal - 1) + '落ち以内の勝率 **' + round1(safeWr) + '%**');
-    }
-
-    results.push({
-      cost: cost,
-      cost_label: COST_LABEL[cost],
-      matches: data.length,
-      fatal_deaths: fatal,
-      fatal_cost: cost * fatal,
-      buckets: buckets,
-      tips: tips,
-    });
-  });
-
-  return results;
-}
-
-function computeSeason(matches) {
-  function getSeason(dateStr) {
-    var parts = dateStr.substring(0, 10).split('-');
-    var year = parseInt(parts[0], 10);
-    var m = parseInt(parts[1], 10);
-    var startMonth, sYear;
-    if (m === 12) {
-      startMonth = 12; sYear = year;
-    } else if (m % 2 === 0) {
-      startMonth = m; sYear = year;
-    } else {
-      if (m === 1) { startMonth = 12; sYear = year - 1; }
-      else { startMonth = m - 1; sYear = year; }
-    }
-    var endMonth = startMonth < 12 ? startMonth + 1 : 1;
-    var endYear = startMonth < 12 ? sYear : sYear + 1;
-    if (startMonth === 12) {
-      return sYear + '年' + startMonth + '月-' + endYear + '年' + endMonth + '月';
-    }
-    return sYear + '年' + startMonth + '-' + endMonth + '月';
-  }
-
-  function getSeasonHalf(dateStr) {
-    var m = parseInt(dateStr.substring(5, 7), 10);
-    return (m === 12 || m % 2 === 0) ? '前半' : '後半';
-  }
-
-  var seasonData = {};
-  var seasonHalf = {};
-
-  matches.forEach(function (d) {
-    var s = getSeason(d.date);
-    var h = getSeasonHalf(d.date);
-    if (!seasonData[s]) seasonData[s] = [];
-    seasonData[s].push(d);
-    if (!seasonHalf[s]) seasonHalf[s] = {};
-    if (!seasonHalf[s][h]) seasonHalf[s][h] = [];
-    seasonHalf[s][h].push(d);
-  });
-
-  var results = [];
-  Object.keys(seasonData).sort().forEach(function (seasonName) {
-    var data = seasonData[seasonName];
-    var firstHalf = (seasonHalf[seasonName] && seasonHalf[seasonName]['前半']) || [];
-    var secondHalf = (seasonHalf[seasonName] && seasonHalf[seasonName]['後半']) || [];
-
-    var tips = [];
-    if (firstHalf.length && secondHalf.length) {
-      var fWr = jsWinRate(firstHalf);
-      var sWr = jsWinRate(secondHalf);
-      var diff = sWr - fWr;
-      if (Math.abs(diff) >= 5) {
-        if (diff > 0) {
-          tips.push('後半が **+' + Math.round(diff) + '%** → シーズン後半に安定');
-        } else {
-          tips.push('前半が **+' + Math.round(-diff) + '%** → 後半は対戦環境が厳しくなった可能性');
-        }
-      }
-    }
-
-    var entry = {
-      name: seasonName,
-      matches: data.length,
-      win_rate: round1(jsWinRate(data)),
-      dmg_efficiency: round3(jsDmgEfficiency(data)),
-      tips: tips,
-    };
-    if (firstHalf.length) {
-      entry.first_half = {
-        matches: firstHalf.length,
-        win_rate: round1(jsWinRate(firstHalf)),
-        dmg_efficiency: round3(jsDmgEfficiency(firstHalf)),
-      };
-    }
-    if (secondHalf.length) {
-      entry.second_half = {
-        matches: secondHalf.length,
-        win_rate: round1(jsWinRate(secondHalf)),
-        dmg_efficiency: round3(jsDmgEfficiency(secondHalf)),
-      };
-    }
-    results.push(entry);
-  });
-
-  return results;
-}
-
-function computeBurstCount(matches) {
-  var byCount = {};
-  matches.forEach(function (d) {
-    if (!d.actions || !d.actions.length) return;
-    var count = d.bursts || 0;
-    if (!byCount[count]) byCount[count] = [];
-    byCount[count].push(d);
-  });
-
-  var countKeys = Object.keys(byCount).map(Number);
-  if (!countKeys.length) return null;
-
-  var results = [];
-  countKeys.sort(function (a, b) { return a - b; }).forEach(function (count) {
-    var bMatches = byCount[count];
-    var wr = jsWinRate(bMatches);
-    results.push({
-      count: count,
-      label: count > 0 ? count + '回' : '0回（未覚醒）',
-      matches: bMatches.length,
-      win_rate: round1(wr),
-    });
-  });
-
-  var tips = [];
-  if (byCount[2] && byCount[2].length >= 3) {
-    var wr2 = jsWinRate(byCount[2]);
-    var others = [];
-    Object.keys(byCount).forEach(function (c) {
-      if (Number(c) < 2) { others = others.concat(byCount[c]); }
-    });
-    if (others.length) {
-      var wrOther = jsWinRate(others);
-      var diff = wr2 - wrOther;
-      if (diff > 0) {
-        tips.push('2回覚醒できた試合の勝率が **' + Math.round(diff) + '%** 高い → ゲージ管理と耐久管理が重要');
-      }
-    }
-  }
-
-  return { by_count: results, tips: tips };
-}
-
-function jsGetDeathEvents(actions) {
-  return (actions || []).filter(function (a) { return a.action === 'death'; });
-}
-function jsGetBurstEvents(actions) {
-  return (actions || []).filter(function (a) { return a.action === 'exbst-f' || a.action === 'exbst-s' || a.action === 'exbst-e'; });
-}
-function jsGetExReadyEvents(actions) {
-  return (actions || []).filter(function (a) { return a.action === 'ex'; });
-}
-
-function computeFallOrder(matches) {
-  var noFall = [], firstFall = [], secondFall = [], sameTime = [];
-  matches.forEach(function (d) {
-    if (!d.actions || !d.actions.length) return;
-    var myDeaths = jsGetDeathEvents(d.actions);
-    var partnerDeaths = jsGetDeathEvents(d.partner_actions);
-    if (!myDeaths.length) {
-      noFall.push(d);
-    } else if (!partnerDeaths.length) {
-      firstFall.push(d);
-    } else {
-      var myFirst = myDeaths[0].action_start_sec;
-      var partnerFirst = partnerDeaths[0].action_start_sec;
-      if (myFirst < partnerFirst) firstFall.push(d);
-      else if (myFirst > partnerFirst) secondFall.push(d);
-      else sameTime.push(d);
-    }
-  });
-  var total = noFall.length + firstFall.length + secondFall.length + sameTime.length;
-  if (total === 0) return null;
-  function buildStats(ms) {
-    var tg = 0, tt = 0;
-    ms.forEach(function (d) { tg += d.dmg_given; tt += d.dmg_taken; });
-    return {
-      count: ms.length,
-      rate: round1(ms.length / total * 100),
-      win_rate: ms.length ? round1(jsWinRate(ms)) : 0,
-      avg_dmg_given: ms.length ? Math.round(jsAvg(ms.map(function (d) { return d.dmg_given; }))) : 0,
-      avg_dmg_taken: ms.length ? Math.round(jsAvg(ms.map(function (d) { return d.dmg_taken; }))) : 0,
-      dmg_efficiency: ms.length ? round3(tt > 0 ? tg / tt : 0) : 0,
-    };
-  }
-  var noFallWr = noFall.length ? jsWinRate(noFall) : 0;
-  var firstWr = firstFall.length ? jsWinRate(firstFall) : 0;
-  var secondWr = secondFall.length ? jsWinRate(secondFall) : 0;
-  var tips = [];
-  if (firstFall.length && secondFall.length) {
-    var diff = secondWr - firstWr;
-    if (Math.abs(diff) >= 5) {
-      var better = diff > 0 ? '後落ち' : '先落ち';
-      tips.push('**' + better + '**の方が勝率 **' + Math.round(Math.abs(diff)) + '%** 高い');
-    }
-  }
-  var fallTotal = firstFall.length + secondFall.length + sameTime.length;
-  if (firstFall.length && fallTotal > 0) {
-    var firstRate = firstFall.length / fallTotal * 100;
-    if (firstRate >= 60) tips.push('先落ち率 **' + Math.round(firstRate) + '%** → 前に出すぎている可能性');
-  }
-  if (noFall.length && firstFall.length) {
-    var d2 = noFallWr - firstWr;
-    if (d2 >= 10) tips.push('0落ちの試合は先落ちより勝率 **' + Math.round(d2) + '%** 高い → 耐久管理が重要');
-  }
-  return {
-    total: total,
-    no_fall: buildStats(noFall),
-    first_fall: buildStats(firstFall),
-    second_fall: buildStats(secondFall),
-    same_time: buildStats(sameTime),
-    tips: tips,
-  };
-}
-
-function computeBurstHoldDeath(matches) {
-  var holdByDeath = {};
-  var noHold = [];
-  var total = 0;
-  matches.forEach(function (d) {
-    if (!d.actions || !d.actions.length) return;
-    var deaths = jsGetDeathEvents(d.actions).sort(function (a, b) { return a.action_start_sec - b.action_start_sec; });
-    var bursts = jsGetBurstEvents(d.actions);
-    var exReadies = jsGetExReadyEvents(d.actions);
-    if (!deaths.length) return;
-    total++;
-    var hasAnyHold = false;
-    deaths.forEach(function (death, i) {
-      var deathTime = death.action_start_sec;
-      var relevantEx = exReadies.filter(function (e) { return e.action_start_sec < deathTime; });
-      if (!relevantEx.length) return;
-      var lastExTime = Math.max.apply(null, relevantEx.map(function (e) { return e.action_start_sec; }));
-      var burstUsed = bursts.some(function (b) {
-        return b.action_start_sec >= lastExTime && b.action_start_sec < deathTime;
-      });
-      if (!burstUsed) {
-        var nth = i + 1;
-        if (!holdByDeath[nth]) holdByDeath[nth] = [];
-        holdByDeath[nth].push(d);
-        hasAnyHold = true;
-      }
-    });
-    if (!hasAnyHold) noHold.push(d);
-  });
-  if (total === 0) return null;
-  function buildStats(ms) {
-    return {
-      count: ms.length,
-      rate: round1(ms.length / total * 100),
-      win_rate: ms.length ? round1(jsWinRate(ms)) : 0,
-    };
-  }
-  var byDeath = [];
-  Object.keys(holdByDeath).map(Number).sort(function (a, b) { return a - b; }).forEach(function (nth) {
-    var stats = buildStats(holdByDeath[nth]);
-    stats.label = nth + '機目に抱え落ち';
-    byDeath.push(stats);
-  });
-  var tips = [];
-  if (byDeath.length && noHold.length) {
-    var holdSet = new Set();
-    Object.keys(holdByDeath).forEach(function (k) {
-      holdByDeath[k].forEach(function (d) { holdSet.add(d); });
-    });
-    var uniqueHold = Array.from(holdSet);
-    var diff = jsWinRate(noHold) - jsWinRate(uniqueHold);
-    if (diff > 0) tips.push('抱え落ちなしの試合の方が勝率 **' + Math.round(diff) + '%** 高い');
-  }
-  return { total: total, by_death: byDeath, no_hold: buildStats(noHold), tips: tips };
-}
-
-function computeFixedPartners(matches, tagPartners) {
-  if (!tagPartners || !tagPartners.length) {
-    return { notice: 'タッグ情報が見つかりませんでした。フレンドを登録してタッグを組むと、固定相方の詳細分析が利用できます。', partners: [] };
-  }
-  var partnerNames = {};
-  tagPartners.forEach(function (tp) { partnerNames[tp.player_name] = tp.team_name; });
-  var fixedMatches = {};
-  matches.forEach(function (d) {
-    if (partnerNames[d.partner_name]) {
-      if (!fixedMatches[d.partner_name]) fixedMatches[d.partner_name] = [];
-      fixedMatches[d.partner_name].push(d);
-    }
-  });
-  var partnerKeys = Object.keys(fixedMatches).sort(function (a, b) { return fixedMatches[b].length - fixedMatches[a].length; });
-  if (!partnerKeys.length) return { partners: [] };
-
-  var results = [];
-  partnerKeys.forEach(function (pName) {
-    var data = fixedMatches[pName];
-    var n = data.length;
-    var wl = jsWinsLosses(data);
-    var w = wl[0], l = wl[1];
-    var wr = w / n * 100;
-
-    var myTg = 0, myTt = 0;
-    data.forEach(function (d) { myTg += d.dmg_given; myTt += d.dmg_taken; });
-    var myEff = myTt > 0 ? myTg / myTt : 0;
-    var myAvgGiven = jsAvg(data.map(function (d) { return d.dmg_given; }));
-    var myAvgTaken = jsAvg(data.map(function (d) { return d.dmg_taken; }));
-    var myTotalKills = 0, myTotalDeaths = 0;
-    data.forEach(function (d) { myTotalKills += d.kills; myTotalDeaths += d.deaths; });
-    var myKd = myTotalDeaths > 0 ? myTotalKills / myTotalDeaths : 0;
-    var myAvgEx = jsAvg(data.map(function (d) { return d.ex_dmg; }));
-    var myBursts = jsAvgBursts(data);
-
-    var pTg = 0, pTt = 0, pTk = 0, pTd = 0;
-    data.forEach(function (d) { pTg += d.partner_dmg_given; pTt += d.partner_dmg_taken; pTk += d.partner_kills; pTd += d.partner_deaths; });
-    var pEff = pTt > 0 ? pTg / pTt : 0;
-    var pAvgGiven = jsAvg(data.map(function (d) { return d.partner_dmg_given; }));
-    var pAvgTaken = jsAvg(data.map(function (d) { return d.partner_dmg_taken; }));
-    var pAvgKills = jsAvg(data.map(function (d) { return d.partner_kills; }));
-    var pAvgDeaths = jsAvg(data.map(function (d) { return d.partner_deaths; }));
-    var pKd = pTd > 0 ? pTk / pTd : 0;
-    var pAvgEx = jsAvg(data.map(function (d) { return d.partner_ex_dmg; }));
-    var pBurstCounts = data.filter(function (d) { return d.partner_actions && d.partner_actions.length; }).map(function (d) { return jsGetBurstEvents(d.partner_actions).length; });
-    var pBursts = pBurstCounts.length ? jsAvg(pBurstCounts) : null;
-
-    var partnerMsMap = {};
-    data.forEach(function (d) {
-      if (!partnerMsMap[d.partner_ms]) partnerMsMap[d.partner_ms] = [];
-      partnerMsMap[d.partner_ms].push(d);
-    });
-    var msBreakdown = [];
-    var msKeys = Object.keys(partnerMsMap);
-    if (msKeys.length > 1 || msKeys.some(function (k) { return partnerMsMap[k].length >= 2; })) {
-      msKeys.sort(function (a, b) { return partnerMsMap[b].length - partnerMsMap[a].length; }).forEach(function (ms) {
-        var msList = partnerMsMap[ms];
-        var msPTg = 0, msPTt = 0;
-        msList.forEach(function (d) { msPTg += d.partner_dmg_given; msPTt += d.partner_dmg_taken; });
-        msBreakdown.push({
-          ms: ms,
-          matches: msList.length,
-          win_rate: round1(jsWinRate(msList)),
-          partner_dmg_efficiency: round3(msPTt > 0 ? msPTg / msPTt : 0),
-        });
-      });
-    }
-
-    var winData = data.filter(function (d) { return d.win; });
-    var lossData = data.filter(function (d) { return !d.win; });
-    function avgOf(key, rows) { return rows.length ? jsAvg(rows.map(function (d) { return d[key]; })) : 0; }
-    function kdOf(kKey, dKey, rows) { var tk = 0, td = 0; rows.forEach(function (d) { tk += d[kKey]; td += d[dKey]; }); return td > 0 ? tk / td : 0; }
-    function effOf(gKey, tKey, rows) { var tg2 = 0, tt2 = 0; rows.forEach(function (d) { tg2 += d[gKey]; tt2 += d[tKey]; }); return tt2 > 0 ? tg2 / tt2 : 0; }
-    function burstsOf(actKey, rows) {
-      var c = rows.filter(function (d) { return d[actKey] && d[actKey].length; }).map(function (d) { return jsGetBurstEvents(d[actKey]).length; });
-      return c.length ? jsAvg(c) : null;
-    }
-    function rnd(v, nd) { if (v == null) return null; var f = Math.pow(10, nd); return Math.round(v * f) / f; }
-    function makeWlMetrics(gKey, tKey, kKey, dKey, exKey, actKey) {
-      var metrics = [];
-      function add(label, wv, lv, nd) { metrics.push({ label: label, win_avg: rnd(wv, nd), loss_avg: rnd(lv, nd) }); }
-      add('平均与ダメージ', avgOf(gKey, winData), avgOf(gKey, lossData), 1);
-      add('平均被ダメージ', avgOf(tKey, winData), avgOf(tKey, lossData), 1);
-      add('与被ダメ比', effOf(gKey, tKey, winData), effOf(gKey, tKey, lossData), 3);
-      add('平均撃墜', avgOf(kKey, winData), avgOf(kKey, lossData), 2);
-      add('平均被撃墜', avgOf(dKey, winData), avgOf(dKey, lossData), 2);
-      add('K/D比', kdOf(kKey, dKey, winData), kdOf(kKey, dKey, lossData), 2);
-      add('平均EXダメージ', avgOf(exKey, winData), avgOf(exKey, lossData), 1);
-      add('平均覚醒回数', burstsOf(actKey, winData), burstsOf(actKey, lossData), 2);
-      return { metrics: metrics };
-    }
-    var myWl = makeWlMetrics('dmg_given', 'dmg_taken', 'kills', 'deaths', 'ex_dmg', 'actions');
-    var partnerWl = makeWlMetrics('partner_dmg_given', 'partner_dmg_taken', 'partner_kills', 'partner_deaths', 'partner_ex_dmg', 'partner_actions');
-
-    var tips = [];
-    if (pEff < 0.8) tips.push('相方の与被ダメ比 **' + pEff.toFixed(3) + '** → カットやライン維持を意識');
-    if (wr < 45 && n >= 5) tips.push('勝率 **' + Math.round(wr) + '%** → 連携や機体の組み合わせを見直し');
-    if (n >= 5) {
-      if (wr >= 90) tips.push('勝率 **' + Math.round(wr) + '%** → 驚異的！全国大会優勝レベル');
-      else if (wr >= 80) tips.push('勝率 **' + Math.round(wr) + '%** → 圧巻！勝ちパターンの再現性を高めよう');
-      else if (wr >= 70) tips.push('勝率 **' + Math.round(wr) + '%** → 素晴らしい相性。この相方を軸に苦手機体の対策を');
-      else if (wr >= 60) tips.push('勝率 **' + Math.round(wr) + '%** → 好調。役割分担を意識してさらに上へ');
-    }
-
-    var entry = {
-      partner_name: pName,
-      matches: n, wins: w, losses: l,
-      win_rate: round1(wr),
-      my_stats: {
-        avg_dmg_given: Math.round(myAvgGiven), avg_dmg_taken: Math.round(myAvgTaken),
-        dmg_efficiency: round3(myEff),
-        avg_kills: round2(jsAvg(data.map(function (d) { return d.kills; }))),
-        avg_deaths: round2(jsAvg(data.map(function (d) { return d.deaths; }))),
-        kd_ratio: round2(myKd), avg_ex_dmg: Math.round(myAvgEx),
-        avg_bursts: myBursts != null ? round2(myBursts) : null,
-      },
-      partner_stats: {
-        avg_dmg_given: Math.round(pAvgGiven), avg_dmg_taken: Math.round(pAvgTaken),
-        dmg_efficiency: round3(pEff),
-        avg_kills: round2(pAvgKills), avg_deaths: round2(pAvgDeaths),
-        kd_ratio: round2(pKd), avg_ex_dmg: Math.round(pAvgEx),
-        avg_bursts: pBursts != null ? round2(pBursts) : null,
-      },
-      my_win_loss_pattern: myWl,
-      partner_win_loss_pattern: partnerWl,
-      partner_ms_breakdown: msBreakdown,
-      tips: tips,
-    };
-    if (partnerNames[pName]) entry.team_name = partnerNames[pName];
-    results.push(entry);
-  });
-  return { partners: results };
-}
-
-// --- Utility ---
-function esc(s) {
-  if (s == null) return '';
-  return String(s).replace(/&/g, '&amp;').replace(/</g, '&lt;').replace(/>/g, '&gt;').replace(/"/g, '&quot;');
-}
-
-// **太字** をパースして html`<strong>太字</strong>` に変換
-function boldText(s) {
-  if (s == null) return '';
-  var parts = String(s).split(/\*\*(.+?)\*\*/g);
-  if (parts.length === 1) return s;
-  return parts.map(function (part, i) {
-    return i % 2 === 1 ? html`<strong class="tip-bold">${part}</strong>` : part;
-  });
-}
-
-function pct(n) { return n != null ? n.toFixed(1) + '%' : '-'; }
-function num(n, d) { return n != null ? n.toFixed(d != null ? d : 0) : '-'; }
-
-// --- Color helpers (4段階: great > good > bad > terrible) ---
-// higherIsBetter: true=値が大きいほど良い, false=値が小さいほど良い
-function valClass4(n, great, good, bad, terrible, higherIsBetter) {
-  if (n == null) return '';
-  if (higherIsBetter) {
-    if (n >= great) return 'val-great';
-    if (n >= good) return 'val-good';
-    if (n <= terrible) return 'val-terrible';
-    return 'val-bad';
-  } else {
-    if (n <= great) return 'val-great';
-    if (n <= good) return 'val-good';
-    if (n >= terrible) return 'val-terrible';
-    return 'val-bad';
-  }
-}
-
-function colorVal(n, great, good, bad, terrible, higherIsBetter, decimals) {
-  if (n == null) return '-';
-  var cls = valClass4(n, great, good, bad, terrible, higherIsBetter);
-  var text = n.toFixed(decimals != null ? decimals : 0);
-  return { sortValue: n, display: html`<span class=${cls}>${text}</span>` };
-}
-
-// 勝率: ≥60=great, ≥50=good, <50=bad, ≤40=terrible
-function colorPct(n) {
-  if (n == null) return '-';
-  var cls = valClass4(n, 60, 50, 50, 40, true);
-  var text = n.toFixed(1) + '%';
-  return { sortValue: n, display: html`<span class=${cls}>${text}</span>` };
-}
-
-// 与被ダメ比: ≥1.2=great, ≥1.0=good, <1.0=bad, ≤0.8=terrible
-function colorDE(n, d) {
-  if (n == null) return '-';
-  var cls = valClass4(n, 1.2, 1.0, 1.0, 0.8, true);
-  var text = n.toFixed(d != null ? d : 3);
-  return { sortValue: n, display: html`<span class=${cls}>${text}</span>` };
-}
-
-// 与ダメ: ≥1100=great, ≥900=good, <900=bad, ≤700=terrible
-function colorDmgGiven(n) { return colorVal(n, 1100, 900, 900, 700, true, 0); }
-// 被ダメ: <700=great, <800=good, ≥800=bad, ≥900=terrible
-function colorDmgTaken(n) { return colorVal(n, 700, 800, 800, 900, false, 0); }
-// 撃墜: ≥1.8=great, ≥1.5=good, <1.5=bad, ≤1.0=terrible
-function colorKills(n) { return colorVal(n, 1.8, 1.5, 1.5, 1.0, true, 2); }
-// 被撃墜: <1.0=great, ≤1.5=good, >1.5=bad, ≥2.5=terrible
-function colorDeaths(n) { return colorVal(n, 1.0, 1.5, 1.5, 2.5, false, 2); }
-// K/D比: ≥1.5=great, ≥1.0=good, <1.0=bad, ≤0.6=terrible
-function colorKD(n) { return colorVal(n, 1.5, 1.0, 1.0, 0.6, true, 2); }
-// EXダメ: ≥200=great, ≥160=good, <160=bad, ≤100=terrible
-function colorExDmg(n) { return colorVal(n, 200, 160, 160, 100, true, 0); }
-// 覚醒回数(最大3程度): ≥2.0=great, ≥1.5=good, <1.5=bad, ≤1.0=terrible
-function colorBursts(n) { return colorVal(n, 2.0, 1.5, 1.5, 1.0, true, 2); }
-// 差分: 色なし
-function colorDiff(n, d) {
-  if (n == null) return '-';
-  var text = (n >= 0 ? '+' : '') + n.toFixed(d != null ? d : 1);
-  return text;
-}
-
-function cellValue(cell) {
-  return cell != null && typeof cell === 'object' && cell.sortValue != null ? cell.sortValue : cell;
-}
-
-function cellDisplay(cell) {
-  return cell != null && typeof cell === 'object' && cell.display != null ? cell.display : cell;
-}
-
-// --- Share helpers ---
-var SVG_X = '<svg viewBox="0 0 24 24"><path d="M18.244 2.25h3.308l-7.227 8.26 8.502 11.24H16.17l-5.214-6.817L4.99 21.75H1.68l7.73-8.835L1.254 2.25H8.08l4.713 6.231zm-1.161 17.52h1.833L7.084 4.126H5.117z"/></svg>';
-var SVG_BSKY = '<svg viewBox="0 0 568 501"><path d="M123.121 33.664C188.241 82.553 258.281 181.68 284 234.873c25.719-53.192 95.759-152.32 160.879-201.21C491.866-1.611 568-28.906 568 57.947c0 17.346-9.945 145.713-15.778 166.555-20.275 72.453-94.155 90.933-159.875 79.748C507.222 323.8 536.444 388.56 473.333 453.32c-119.86 122.992-172.272-30.859-185.702-70.281-2.462-7.227-3.614-10.608-3.631-7.733-.017-2.875-1.169.506-3.631 7.733-13.43 39.422-65.842 193.273-185.702 70.281-63.111-64.76-33.889-129.52 80.986-149.071-65.72 11.185-139.6-7.295-159.875-79.748C9.945 203.659 0 75.291 0 57.946 0-28.906 76.135-1.612 123.121 33.664z"/></svg>';
-var SVG_LINE = '<svg viewBox="0 0 24 24"><path d="M19.365 9.863c.349 0 .63.285.63.631 0 .345-.281.63-.63.63H17.61v1.125h1.755c.349 0 .63.283.63.63 0 .344-.281.629-.63.629h-2.386c-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63h2.386c.346 0 .627.285.627.63 0 .349-.281.63-.63.63H17.61v1.125h1.755zm-3.855 3.016c0 .27-.174.51-.432.596-.064.021-.133.031-.199.031-.211 0-.391-.09-.51-.25l-2.443-3.317v2.94c0 .344-.279.629-.631.629-.346 0-.626-.285-.626-.629V8.108c0-.27.173-.51.43-.595.06-.023.136-.033.194-.033.195 0 .375.104.495.254l2.462 3.33V8.108c0-.345.282-.63.63-.63.345 0 .63.285.63.63v4.771zm-5.741 0c0 .344-.282.629-.631.629-.345 0-.627-.285-.627-.629V8.108c0-.345.282-.63.63-.63.346 0 .628.285.628.63v4.771zm-2.466.629H4.917c-.345 0-.63-.285-.63-.629V8.108c0-.345.285-.63.63-.63.348 0 .63.285.63.63v4.141h1.756c.348 0 .629.283.629.63 0 .344-.282.629-.629.629M24 10.314C24 4.943 18.615.572 12 .572S0 4.943 0 10.314c0 4.811 4.27 8.842 10.035 9.608.391.082.923.258 1.058.59.12.301.079.766.038 1.08l-.164 1.02c-.045.301-.24 1.186 1.049.645 1.291-.539 6.916-4.078 9.436-6.975C23.176 14.393 24 12.458 24 10.314"/></svg>';
-var SVG_COPY = '<svg viewBox="0 0 24 24"><path d="M16 1H4c-1.1 0-2 .9-2 2v14h2V3h12V1zm3 4H8c-1.1 0-2 .9-2 2v14c0 1.1.9 2 2 2h11c1.1 0 2-.9 2-2V7c0-1.1-.9-2-2-2zm0 16H8V7h11v14z"/></svg>';
-var SVG_CHECK = '<svg viewBox="0 0 24 24"><path d="M9 16.17L4.83 12l-1.42 1.41L9 19 21 7l-1.41-1.41z"/></svg>';
-
-function buildShareText(items) {
-  var lines = ['【EXVS2IB 戦績診断】'];
-  for (var i = 0; i < items.length; i++) {
-    var item = items[i];
-    if (item.type === 'top_ms') {
-      lines.push('🤖 最多使用: ' + item.ms + '（' + item.count + '戦）');
-    } else if (item.type === 'strong_enemy') {
-      lines.push('💪 ' + item.enemy + '相手に勝率' + item.wr + '%！');
-    } else if (item.type === 'weak_enemy') {
-      lines.push('😈 ' + item.enemy + 'に勝率' + item.wr + '%...天敵かも');
-    } else if (item.type === 'dmg_efficiency') {
-      var desc = item.value >= 1.0 ? '与ダメが上回ってます' : '被ダメが上回ってます';
-      lines.push('⚔ ' + item.ms + 'の与被ダメ比: ' + item.value + '（' + desc + '）');
-    }
-  }
-  lines.push('');
-  lines.push('▶ 自分も診断してみる');
-  lines.push(location.origin);
-  return lines.join('\n');
-}
-
-// --- Generic components ---
-
-function Tips({ tips }) {
-  if (!tips || !tips.length) return null;
-  return html`<blockquote><strong>💡アドバイス:</strong><br />${tips.map(function (t, i) {
-    var text = typeof t === 'string' ? t : t.text;
-    var details = typeof t === 'object' && t.details ? t.details : null;
-    return html`${i > 0 && html`<br />`}${boldText(text)}
-      ${details && html`<ul class="advice-details">${details.map(function (d) { return html`<li>${boldText(d)}</li>`; })}</ul>`}`;
-  })}</blockquote>`;
-}
-
-function SortableTable({ headers, rows, sortableColumns, defaultLimit }) {
-  if (!rows || !rows.length) return null;
-  var sortRef = useState({ col: -1, asc: true });
-  var sortState = sortRef[0], setSortState = sortRef[1];
-  var limitRef = useState(defaultLimit || 0);
-  var limit = limitRef[0], setLimit = limitRef[1];
-
-  var sortedRows = useMemo(function () {
-    if (sortState.col < 0) return rows;
-    var col = sortState.col;
-    var sorted = rows.slice().sort(function (a, b) {
-      var va = cellValue(a[col]), vb = cellValue(b[col]);
-      // 数値文字列からパース（%や+を除去）
-      var na = typeof va === 'number' ? va : parseFloat(String(va).replace(/[%+戦件回]/g, ''));
-      var nb = typeof vb === 'number' ? vb : parseFloat(String(vb).replace(/[%+戦件回]/g, ''));
-      if (!isNaN(na) && !isNaN(nb)) {
-        return sortState.asc ? na - nb : nb - na;
-      }
-      var sa = String(cellValue(a[col])), sb = String(cellValue(b[col]));
-      return sortState.asc ? sa.localeCompare(sb) : sb.localeCompare(sa);
-    });
-    return sorted;
-  }, [rows, sortState]);
-
-  var expanded = limit === 0;
-  var displayRows = limit > 0 ? sortedRows.slice(0, limit) : sortedRows;
-  var hasMore = limit > 0 && sortedRows.length > limit;
-
-  function handleSort(colIdx) {
-    if (sortState.col === colIdx) {
-      if (colIdx === 0 && !sortState.asc) {
-        // 1列目で降順→リセット（元の順序に戻す）
-        setSortState({ col: -1, asc: true });
-      } else {
-        setSortState({ col: colIdx, asc: !sortState.asc });
-      }
-    } else {
-      setSortState({ col: colIdx, asc: colIdx === 0 ? true : false });
-    }
-  }
-
-  var sortable = sortableColumns || [];
-
-  return html`<div>
-    <div class="table-wrap"><table>
-      <thead><tr>${headers.map(function (h, i) {
-        var isSortable = h !== '' && (sortable.length === 0 || sortable.indexOf(i) >= 0);
-        var indicator = sortState.col === i ? (sortState.asc ? ' ▲' : ' ▼') : (isSortable ? ' ▽' : '');
-        return html`<th class=${isSortable ? 'sortable' : ''} onClick=${isSortable ? function () { handleSort(i); } : undefined}>${h}${indicator}</th>`;
-      })}</tr></thead>
-      <tbody>${displayRows.map(function (row) {
-        return html`<tr>${row.map(function (cell) { return html`<td>${cellDisplay(cell)}</td>`; })}</tr>`;
-      })}</tbody>
-    </table></div>
-    ${defaultLimit > 0 && sortedRows.length > defaultLimit && html`<div class="show-more-wrap">
-      ${hasMore
-        ? html`<button class="show-more-btn" onClick=${function () { setLimit(0); }}>もっと見る (+${sortedRows.length - limit}件)</button>`
-        : html`<button class="show-more-btn" onClick=${function () { setLimit(defaultLimit); }}>折りたたむ</button>`}
-    </div>`}
-  </div>`;
-}
-
-function Table({ headers, rows }) {
-  if (!rows || !rows.length) return null;
-  return html`<${SortableTable} headers=${headers} rows=${rows} />`;
-}
-
-function Section({ title, open, children }) {
-  return html`<details ...${{ open: open || false }}>
-    <summary><strong>${title}</strong></summary>
-    ${children}
-  </details><hr />`;
-}
-
-function SubSection({ title, open, children }) {
-  return html`<details ...${{ open: open || false }}>
-    <summary>${title}</summary>
-    ${children}
-  </details>`;
-}
 
 // --- Calendar component ---
 
@@ -1487,748 +292,6 @@ function PeriodSelector({ periods, selected, onSelect, userKey, onCustomReport }
       </div>`}
     </div>`}
   </div>`;
-}
-
-// --- Report sections ---
-
-function EnemyMatchupSection({ matchup }) {
-  if (!matchup) return null;
-  var headers = ['機体名', '試合', '勝率', '与被ダメ比', '与ダメ', '被ダメ'];
-  function matchupRows(list) {
-    return (list || []).map(function (e) {
-      return [esc(e.ms), e.matches, colorPct(e.win_rate), colorDE(e.dmg_efficiency, 3), colorDmgGiven(e.avg_dmg_given), colorDmgTaken(e.avg_dmg_taken)];
-    });
-  }
-  return html`<div>
-    ${matchup.strong && matchup.strong.length > 0 && html`<p><strong>得意な相手:</strong></p><${SortableTable} headers=${headers} rows=${matchupRows(matchup.strong)} defaultLimit=${5} />`}
-    ${matchup.weak && matchup.weak.length > 0 && html`<p><strong>苦手な相手:</strong></p><${SortableTable} headers=${headers} rows=${matchupRows(matchup.weak)} defaultLimit=${5} />`}
-    ${matchup.even && matchup.even.length > 0 && html`<p><strong>互角の相手:</strong></p><${SortableTable} headers=${headers} rows=${matchupRows(matchup.even)} defaultLimit=${5} />`}
-    <${Tips} tips=${matchup.tips} />
-  </div>`;
-}
-
-function PartnerSection({ partners }) {
-  if (!partners || !partners.length) return null;
-  var rows = partners.map(function (p) {
-    return [esc(p.ms), p.matches, colorPct(p.win_rate), colorDE(p.dmg_efficiency, 3)];
-  });
-  return html`<div>
-    <${SortableTable} headers=${['機体名', '試合', '勝率', '与被ダメ比']} rows=${rows} defaultLimit=${10} />
-  </div>`;
-}
-
-
-function MsPairSubSection({ msPair }) {
-  if (!msPair) return null;
-  var list = msPair.by_matches || [];
-  if (!list.length) return null;
-  var rows = list.map(function (p) {
-    return [esc(p.pair), p.matches, colorPct(p.win_rate), colorDE(p.dmg_efficiency, 3)];
-  });
-  return html`<div>
-    <${SortableTable} headers=${['編成', '試合数', '勝率', '与被ダメ比']} rows=${rows} defaultLimit=${10} />
-  </div>`;
-}
-
-function CostPairSubSection({ costPair }) {
-  if (!costPair || !costPair.length) return null;
-  var rows = costPair.map(function (p) {
-    return [esc(p.pair), p.matches, colorPct(p.win_rate), colorDE(p.dmg_efficiency, 3)];
-  });
-  return html`<div>
-    <${SortableTable} headers=${['コスト編成', '試合数', '勝率', '与被ダメ比']} rows=${rows} defaultLimit=${10} />
-  </div>`;
-}
-
-function DmgContributionSubSection({ dmg }) {
-  if (!dmg) return null;
-  function diffPct(win, lose) {
-    if (win == null || lose == null) return '-';
-    var d = win - lose;
-    var s = d >= 0 ? '+' : '';
-    return s + d.toFixed(1) + '%';
-  }
-  var rows = [];
-  (dmg.by_cost || []).forEach(function (c) {
-    rows.push([c.matches, pct(c.avg_contribution), pct(c.avg_win_contribution), pct(c.avg_lose_contribution), diffPct(c.avg_win_contribution, c.avg_lose_contribution)]);
-  });
-  return html`<div>
-    <${Table} headers=${['試合数', '平均貢献率', '勝利時', '敗北時', '差分']} rows=${rows} />
-  </div>`;
-}
-
-function DeathsImpactSubSection({ deaths }) {
-  if (!deaths || !deaths.length) return null;
-  return deaths.map(function (d) {
-    var rows = (d.buckets || []).map(function (b) {
-      return [b.label, b.matches + '戦', colorPct(b.win_rate)];
-    });
-    return html`<div>
-      <${Table} headers=${['被撃墜数', '試合数', '勝率']} rows=${rows} />
-      <${Tips} tips=${d.tips} />
-    </div>`;
-  });
-}
-
-// スクロール連動: 要素が画面に入ったらtrueを返すフック
-function useInView(ref) {
-  var state = useState(false);
-  var inView = state[0], setInView = state[1];
-
-  useEffect(function () {
-    if (!ref.current) return;
-    var observer = new IntersectionObserver(function (entries) {
-      if (entries[0].isIntersecting) {
-        setInView(true);
-        observer.disconnect();
-      }
-    }, { threshold: 0.1 });
-    observer.observe(ref.current);
-    return function () { observer.disconnect(); };
-  }, []);
-
-  return inView;
-}
-
-// 50%基準線プラグイン
-var winRate50Plugin = {
-  id: 'winRate50Line',
-  afterDraw: function (chart) {
-    var yScale = chart.scales.y;
-    if (!yScale) return;
-    var y = yScale.getPixelForValue(50);
-    var ctx = chart.ctx;
-    ctx.save();
-    ctx.beginPath();
-    ctx.setLineDash([6, 4]);
-    ctx.strokeStyle = 'rgba(255, 255, 255, 0.3)';
-    ctx.lineWidth = 1;
-    ctx.moveTo(chart.chartArea.left, y);
-    ctx.lineTo(chart.chartArea.right, y);
-    ctx.stroke();
-    ctx.fillStyle = 'rgba(255, 255, 255, 0.4)';
-    ctx.font = '11px sans-serif';
-    ctx.fillText('50%', chart.chartArea.left + 4, y - 4);
-    ctx.restore();
-  },
-};
-
-function TimeOfDayChart({ hours }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !hours || !hours.length) return;
-    if (chartRef.current) chartRef.current.destroy();
-
-    var labels = hours.map(function (h) { return h.hour + '時'; });
-    var winRates = hours.map(function (h) { return h.win_rate; });
-    var matches = hours.map(function (h) { return h.matches; });
-
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: {
-        labels: labels,
-        datasets: [
-          {
-            label: '勝率 (%)',
-            data: winRates,
-            backgroundColor: winRates.map(function (v) { return v >= 60 ? 'rgba(76, 175, 80, 0.7)' : v < 50 ? 'rgba(239, 83, 80, 0.7)' : 'rgba(129, 212, 250, 0.3)'; }),
-            borderWidth: 0,
-            yAxisID: 'y',
-          },
-          {
-            label: '試合数',
-            data: matches,
-            type: 'line',
-            borderColor: '#81d4fa',
-            backgroundColor: 'rgba(129, 212, 250, 0.1)',
-            fill: false,
-            tension: 0.3,
-            pointRadius: 4,
-            pointHoverRadius: 6,
-            yAxisID: 'y1',
-          },
-        ],
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        interaction: { mode: 'index', intersect: false },
-        plugins: {
-          legend: { labels: { color: '#aaa', font: { size: 12 }, usePointStyle: true, generateLabels: function (chart) { return chart.data.datasets.map(function (ds, i) { var meta = chart.getDatasetMeta(i); var ps; if (ds.type === 'line') { var c = document.createElement('canvas'); c.width = 24; c.height = 12; var cx = c.getContext('2d'); var color = ds.borderColor; cx.strokeStyle = color; cx.lineWidth = 2; cx.beginPath(); cx.moveTo(4, 6); cx.lineTo(20, 6); cx.stroke(); cx.fillStyle = color; cx.beginPath(); cx.arc(4, 6, 3, 0, Math.PI * 2); cx.fill(); cx.beginPath(); cx.arc(20, 6, 3, 0, Math.PI * 2); cx.fill(); ps = c; } else { ps = 'rectRounded'; } return { text: ds.label, fontColor: '#aaa', fillStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.backgroundColor) ? ds.backgroundColor[0] : ds.backgroundColor), strokeStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.borderColor) ? ds.borderColor[0] : ds.borderColor), lineWidth: ds.type === 'line' ? 0 : 1, pointStyle: ps, hidden: meta.hidden, datasetIndex: i }; }); } } },
-        },
-        scales: {
-          x: {
-            ticks: { color: '#888', font: { size: 11 } },
-            grid: { color: 'rgba(255,255,255,0.05)' },
-          },
-          y: {
-            position: 'left',
-            min: 0,
-            max: 100,
-            ticks: { color: '#aaa', callback: function (v) { return v + '%'; } },
-            grid: { color: 'rgba(255,255,255,0.08)' },
-          },
-          y1: {
-            position: 'right',
-            min: 0,
-            ticks: { color: '#aaa', stepSize: 1 },
-            grid: { display: false },
-          },
-        },
-      },
-      plugins: [winRate50Plugin],
-    });
-
-    return function () { if (chartRef.current) { chartRef.current.destroy(); chartRef.current = null; } };
-  }, [hours, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}><canvas ref=${canvasRef} /></div>`;
-}
-
-function DayOfWeekChart({ days }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !days || !days.length) return;
-    if (chartRef.current) chartRef.current.destroy();
-
-    var labels = days.map(function (d) { return d.name + '曜'; });
-    var winRates = days.map(function (d) { return d.win_rate; });
-    var matches = days.map(function (d) { return d.matches; });
-
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: {
-        labels: labels,
-        datasets: [
-          {
-            label: '勝率 (%)',
-            data: winRates,
-            backgroundColor: winRates.map(function (v) { return v >= 60 ? 'rgba(76, 175, 80, 0.7)' : v < 50 ? 'rgba(239, 83, 80, 0.7)' : 'rgba(129, 212, 250, 0.3)'; }),
-            borderWidth: 0,
-            yAxisID: 'y',
-          },
-          {
-            label: '試合数',
-            data: matches,
-            type: 'line',
-            borderColor: '#81d4fa',
-            backgroundColor: 'rgba(129, 212, 250, 0.1)',
-            fill: false,
-            tension: 0.3,
-            pointRadius: 4,
-            pointHoverRadius: 6,
-            yAxisID: 'y1',
-          },
-        ],
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        interaction: { mode: 'index', intersect: false },
-        plugins: {
-          legend: { labels: { color: '#aaa', font: { size: 12 }, usePointStyle: true, generateLabels: function (chart) { return chart.data.datasets.map(function (ds, i) { var meta = chart.getDatasetMeta(i); var ps; if (ds.type === 'line') { var c = document.createElement('canvas'); c.width = 24; c.height = 12; var cx = c.getContext('2d'); var color = ds.borderColor; cx.strokeStyle = color; cx.lineWidth = 2; cx.beginPath(); cx.moveTo(4, 6); cx.lineTo(20, 6); cx.stroke(); cx.fillStyle = color; cx.beginPath(); cx.arc(4, 6, 3, 0, Math.PI * 2); cx.fill(); cx.beginPath(); cx.arc(20, 6, 3, 0, Math.PI * 2); cx.fill(); ps = c; } else { ps = 'rectRounded'; } return { text: ds.label, fontColor: '#aaa', fillStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.backgroundColor) ? ds.backgroundColor[0] : ds.backgroundColor), strokeStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.borderColor) ? ds.borderColor[0] : ds.borderColor), lineWidth: ds.type === 'line' ? 0 : 1, pointStyle: ps, hidden: meta.hidden, datasetIndex: i }; }); } } },
-        },
-        scales: {
-          x: {
-            ticks: { color: '#888', font: { size: 11 } },
-            grid: { color: 'rgba(255,255,255,0.05)' },
-          },
-          y: {
-            position: 'left',
-            min: 0,
-            max: 100,
-            ticks: { color: '#aaa', callback: function (v) { return v + '%'; } },
-            grid: { color: 'rgba(255,255,255,0.08)' },
-          },
-          y1: {
-            position: 'right',
-            min: 0,
-            ticks: { color: '#aaa', stepSize: 1 },
-            grid: { display: false },
-          },
-        },
-      },
-      plugins: [winRate50Plugin],
-    });
-
-    return function () { if (chartRef.current) { chartRef.current.destroy(); chartRef.current = null; } };
-  }, [days, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}><canvas ref=${canvasRef} /></div>`;
-}
-
-function DailyTrendChart({ days }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !days || !days.length) return;
-
-    if (chartRef.current) {
-      chartRef.current.destroy();
-    }
-
-    var labels = days.map(function (d) { return d.date.slice(5); });
-    var winRates = days.map(function (d) { return d.win_rate; });
-    var matches = days.map(function (d) { return d.matches; });
-
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: {
-        labels: labels,
-        datasets: [
-          {
-            label: '勝率 (%)',
-            data: winRates,
-            backgroundColor: winRates.map(function (v) { return v >= 60 ? 'rgba(76, 175, 80, 0.7)' : v < 50 ? 'rgba(239, 83, 80, 0.7)' : 'rgba(129, 212, 250, 0.3)'; }),
-            borderWidth: 0,
-            yAxisID: 'y',
-          },
-          {
-            label: '試合数',
-            data: matches,
-            type: 'line',
-            borderColor: '#81d4fa',
-            backgroundColor: 'rgba(129, 212, 250, 0.1)',
-            fill: false,
-            tension: 0.3,
-            pointRadius: days.length > 30 ? 2 : 4,
-            pointHoverRadius: 6,
-            yAxisID: 'y1',
-          },
-        ],
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        interaction: { mode: 'index', intersect: false },
-        plugins: {
-          legend: { labels: { color: '#aaa', font: { size: 12 }, usePointStyle: true, generateLabels: function (chart) { return chart.data.datasets.map(function (ds, i) { var meta = chart.getDatasetMeta(i); var ps; if (ds.type === 'line') { var c = document.createElement('canvas'); c.width = 24; c.height = 12; var cx = c.getContext('2d'); var color = ds.borderColor; cx.strokeStyle = color; cx.lineWidth = 2; cx.beginPath(); cx.moveTo(4, 6); cx.lineTo(20, 6); cx.stroke(); cx.fillStyle = color; cx.beginPath(); cx.arc(4, 6, 3, 0, Math.PI * 2); cx.fill(); cx.beginPath(); cx.arc(20, 6, 3, 0, Math.PI * 2); cx.fill(); ps = c; } else { ps = 'rectRounded'; } return { text: ds.label, fontColor: '#aaa', fillStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.backgroundColor) ? ds.backgroundColor[0] : ds.backgroundColor), strokeStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.borderColor) ? ds.borderColor[0] : ds.borderColor), lineWidth: ds.type === 'line' ? 0 : 1, pointStyle: ps, hidden: meta.hidden, datasetIndex: i }; }); } } },
-          tooltip: {
-            callbacks: {
-              title: function (items) {
-                var idx = items[0].dataIndex;
-                var d = days[idx];
-                return d.date + ' (' + d.dow_name + ')';
-              },
-            },
-          },
-        },
-        scales: {
-          x: {
-            ticks: { color: '#888', maxRotation: 45, font: { size: 11 } },
-            grid: { color: 'rgba(255,255,255,0.05)' },
-          },
-          y: {
-            position: 'left',
-            min: 0,
-            max: 100,
-            ticks: {
-              color: '#aaa',
-              callback: function (v) { return v + '%'; },
-            },
-            grid: { color: 'rgba(255,255,255,0.08)' },
-          },
-          y1: {
-            position: 'right',
-            min: 0,
-            ticks: { color: '#aaa', stepSize: 1 },
-            grid: { display: false },
-          },
-        },
-      },
-      plugins: [winRate50Plugin],
-    });
-
-    return function () {
-      if (chartRef.current) {
-        chartRef.current.destroy();
-        chartRef.current = null;
-      }
-    };
-  }, [days, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}>
-    <canvas ref=${canvasRef} />
-  </div>`;
-}
-
-// シーズンごとの勝率(棒)と試合数(折れ線)の推移（時間帯/曜日グラフと同デザイン）
-function SeasonChart({ seasons }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !seasons || !seasons.length) return;
-    if (chartRef.current) chartRef.current.destroy();
-
-    // 年が変わった時だけ年を表示する（例: 2026年4-5月 → 6-7月 → 8-9月）
-    // 斜め表示を避けるため、年を含む長いラベルは「年」の位置で2行に分割する
-    var prevYear = null;
-    var labels = seasons.map(function (s) {
-      var m = s.name.match(/^(\d{4})年/);
-      var label = (m && m[1] === prevYear) ? s.name.replace(/^\d{4}年/, '') : s.name;
-      if (m) prevYear = m[1];
-      var yi = label.indexOf('年');
-      return yi === -1 ? label : [label.slice(0, yi + 1), label.slice(yi + 1)];
-    });
-    var winRates = seasons.map(function (s) { return s.win_rate; });
-    var matches = seasons.map(function (s) { return s.matches; });
-
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: {
-        labels: labels,
-        datasets: [
-          {
-            label: '勝率 (%)',
-            data: winRates,
-            backgroundColor: winRates.map(function (v) { return v >= 60 ? 'rgba(76, 175, 80, 0.7)' : v < 50 ? 'rgba(239, 83, 80, 0.7)' : 'rgba(129, 212, 250, 0.3)'; }),
-            borderWidth: 0,
-            yAxisID: 'y',
-          },
-          {
-            label: '試合数',
-            data: matches,
-            type: 'line',
-            borderColor: '#81d4fa',
-            backgroundColor: 'rgba(129, 212, 250, 0.1)',
-            fill: false,
-            tension: 0.3,
-            pointRadius: 4,
-            pointHoverRadius: 6,
-            yAxisID: 'y1',
-          },
-        ],
-      },
-      options: {
-        responsive: true,
-        maintainAspectRatio: false,
-        interaction: { mode: 'index', intersect: false },
-        plugins: {
-          legend: { labels: { color: '#aaa', font: { size: 12 }, usePointStyle: true, generateLabels: function (chart) { return chart.data.datasets.map(function (ds, i) { var meta = chart.getDatasetMeta(i); var ps; if (ds.type === 'line') { var c = document.createElement('canvas'); c.width = 24; c.height = 12; var cx = c.getContext('2d'); var color = ds.borderColor; cx.strokeStyle = color; cx.lineWidth = 2; cx.beginPath(); cx.moveTo(4, 6); cx.lineTo(20, 6); cx.stroke(); cx.fillStyle = color; cx.beginPath(); cx.arc(4, 6, 3, 0, Math.PI * 2); cx.fill(); cx.beginPath(); cx.arc(20, 6, 3, 0, Math.PI * 2); cx.fill(); ps = c; } else { ps = 'rectRounded'; } return { text: ds.label, fontColor: '#aaa', fillStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.backgroundColor) ? ds.backgroundColor[0] : ds.backgroundColor), strokeStyle: ds.type === 'line' ? ds.borderColor : (Array.isArray(ds.borderColor) ? ds.borderColor[0] : ds.borderColor), lineWidth: ds.type === 'line' ? 0 : 1, pointStyle: ps, hidden: meta.hidden, datasetIndex: i }; }); } } },
-          tooltip: {
-            callbacks: {
-              title: function (items) { return seasons[items[0].dataIndex].name; },
-            },
-          },
-        },
-        scales: {
-          x: {
-            ticks: { color: '#888', maxRotation: 0, minRotation: 0, font: { size: 11 } },
-            grid: { color: 'rgba(255,255,255,0.05)' },
-          },
-          y: {
-            position: 'left',
-            min: 0,
-            max: 100,
-            ticks: { color: '#aaa', callback: function (v) { return v + '%'; } },
-            grid: { color: 'rgba(255,255,255,0.08)' },
-          },
-          y1: {
-            position: 'right',
-            min: 0,
-            ticks: { color: '#aaa', stepSize: 1 },
-            grid: { display: false },
-          },
-        },
-      },
-      plugins: [winRate50Plugin],
-    });
-
-    return function () { if (chartRef.current) { chartRef.current.destroy(); chartRef.current = null; } };
-  }, [seasons, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}><canvas ref=${canvasRef} /></div>`;
-}
-
-// --- Reusable win-rate bar chart (categories + win rate + match count) ---
-
-function WinRateBarChart({ items }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !items || !items.length) return;
-    if (chartRef.current) chartRef.current.destroy();
-    var labels = items.map(function (i) { return i.label; });
-    var rates = items.map(function (i) { return i.win_rate; });
-    var counts = items.map(function (i) { return i.matches; });
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: {
-        labels: labels,
-        datasets: [
-          { label: '勝率 (%)', data: rates, backgroundColor: rates.map(function (v) { return v >= 60 ? 'rgba(76, 175, 80, 0.7)' : v < 50 ? 'rgba(239, 83, 80, 0.7)' : 'rgba(129, 212, 250, 0.3)'; }), borderWidth: 0, yAxisID: 'y' },
-          { label: '試合数', data: counts, type: 'line', borderColor: '#81d4fa', fill: false, tension: 0.3, pointRadius: 4, pointHoverRadius: 6, yAxisID: 'y1' },
-        ],
-      },
-      options: {
-        responsive: true, maintainAspectRatio: false,
-        interaction: { mode: 'index', intersect: false },
-        plugins: { legend: { labels: { color: '#aaa', font: { size: 12 } } } },
-        scales: {
-          x: { ticks: { color: '#888', font: { size: 11 } }, grid: { color: 'rgba(255,255,255,0.05)' } },
-          y: { position: 'left', min: 0, max: 100, ticks: { color: '#aaa', callback: function (v) { return v + '%'; } }, grid: { color: 'rgba(255,255,255,0.08)' } },
-          y1: { position: 'right', min: 0, ticks: { color: '#aaa', stepSize: 1 }, grid: { display: false } },
-        },
-      },
-      plugins: [winRate50Plugin],
-    });
-    return function () { if (chartRef.current) { chartRef.current.destroy(); chartRef.current = null; } };
-  }, [items, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}><canvas ref=${canvasRef} /></div>`;
-}
-
-// 勝利時/敗北時のメトリクスを並べたグループ棒グラフ
-// ダメージ貢献率チャート（全体/勝利時/敗北時）
-function DmgContributionChart({ dmg }) {
-  var containerRef = useRef(null);
-  var canvasRef = useRef(null);
-  var chartRef = useRef(null);
-  var inView = useInView(containerRef);
-
-  useEffect(function () {
-    if (!inView || !canvasRef.current || !dmg || !dmg.by_cost || !dmg.by_cost.length) return;
-    if (chartRef.current) chartRef.current.destroy();
-    var c = dmg.by_cost[0];
-    var labels = ['全体', '勝利時', '敗北時'];
-    var values = [c.avg_contribution || 0, c.avg_win_contribution || 0, c.avg_lose_contribution || 0];
-    var colors = ['rgba(129, 212, 250, 0.5)', 'rgba(105, 240, 174, 0.6)', 'rgba(239, 83, 80, 0.6)'];
-    chartRef.current = new Chart(canvasRef.current, {
-      type: 'bar',
-      data: { labels: labels, datasets: [{ label: '貢献率 (%)', data: values, backgroundColor: colors, borderWidth: 0 }] },
-      options: {
-        responsive: true, maintainAspectRatio: false,
-        plugins: { legend: { display: false }, tooltip: { callbacks: { label: function (ctx) { return '貢献率: ' + ctx.parsed.y.toFixed(1) + '%'; } } } },
-        scales: {
-          x: { ticks: { color: '#888' }, grid: { color: 'rgba(255,255,255,0.05)' } },
-          y: { min: 0, max: 100, ticks: { color: '#aaa', callback: function (v) { return v + '%'; } }, grid: { color: 'rgba(255,255,255,0.08)' } },
-        },
-      },
-    });
-    return function () { if (chartRef.current) { chartRef.current.destroy(); chartRef.current = null; } };
-  }, [dmg, inView]);
-
-  return html`<div class="chart-container" ref=${containerRef}><canvas ref=${canvasRef} /></div>`;
-}
-
-// --- Fall order / Burst before death ---
-
-function FallOrderContent({ fallOrder }) {
-  if (!fallOrder) return null;
-  var n = fallOrder.no_fall;
-  var f = fallOrder.first_fall;
-  var s = fallOrder.second_fall;
-  var st = fallOrder.same_time;
-  var rows = [
-    ['0落ち', n.count + '戦 (' + n.rate + '%)', colorPct(n.win_rate), colorDmgGiven(n.avg_dmg_given), colorDmgTaken(n.avg_dmg_taken), colorDE(n.dmg_efficiency, 3)],
-    ['先落ち', f.count + '戦 (' + f.rate + '%)', colorPct(f.win_rate), colorDmgGiven(f.avg_dmg_given), colorDmgTaken(f.avg_dmg_taken), colorDE(f.dmg_efficiency, 3)],
-    ['後落ち', s.count + '戦 (' + s.rate + '%)', colorPct(s.win_rate), colorDmgGiven(s.avg_dmg_given), colorDmgTaken(s.avg_dmg_taken), colorDE(s.dmg_efficiency, 3)],
-  ];
-  if (st.count > 0) {
-    rows.push(['同時落ち', st.count + '戦 (' + st.rate + '%)', colorPct(st.win_rate), colorDmgGiven(st.avg_dmg_given), colorDmgTaken(st.avg_dmg_taken), colorDE(st.dmg_efficiency, 3)]);
-  }
-  return html`<div>
-    <p>対象: ${fallOrder.total}戦</p>
-    <${Table} headers=${['パターン', '試合数', '勝率', '与ダメ', '被ダメ', '与被ダメ比']} rows=${rows} />
-    <${Tips} tips=${fallOrder.tips} />
-  </div>`;
-}
-
-function BurstHoldDeathContent({ holdData }) {
-  if (!holdData) return null;
-  var nh = holdData.no_hold;
-  var rows = (holdData.by_death || []).filter(function (d) { return d.count > 0; }).map(function (d) {
-    return [d.label, d.count + '戦 (' + d.rate + '%)', colorPct(d.win_rate)];
-  });
-  rows.push(['抱え落ちなし', nh.count + '戦 (' + nh.rate + '%)', colorPct(nh.win_rate)]);
-  return html`<div>
-    <p>覚醒ゲージが溜まった状態で発動せずに撃墜された試合（対象: ${holdData.total}戦）</p>
-    <${Table} headers=${['パターン', '試合数', '勝率']} rows=${rows} />
-    <${Tips} tips=${holdData.tips} />
-  </div>`;
-}
-
-function BurstCountContent({ countData }) {
-  if (!countData || !countData.by_count || !countData.by_count.length) return null;
-  var rows = countData.by_count.map(function (c) {
-    return [c.label, c.matches + '戦', colorPct(c.win_rate)];
-  });
-  return html`<div>
-    <${Table} headers=${['覚醒回数', '試合数', '勝率']} rows=${rows} />
-    <${Tips} tips=${countData.tips} />
-  </div>`;
-}
-
-// --- Aggregation helpers (for 全機体 view) ---
-
-function aggregateDeathsImpact(msStats) {
-  var bucketMap = {};
-  Object.keys(msStats).forEach(function (ms) {
-    var di = msStats[ms].deaths_impact;
-    if (!di) return;
-    di.forEach(function (d) {
-      (d.buckets || []).forEach(function (b) {
-        if (!bucketMap[b.label]) bucketMap[b.label] = { label: b.label, matches: 0, wins: 0 };
-        bucketMap[b.label].matches += b.matches;
-        bucketMap[b.label].wins += Math.round(b.win_rate / 100 * b.matches);
-      });
-    });
-  });
-  var buckets = Object.values(bucketMap).map(function (b) {
-    return { label: b.label, matches: b.matches, win_rate: b.matches > 0 ? b.wins / b.matches * 100 : 0 };
-  });
-  if (!buckets.length) return null;
-  return [{ buckets: buckets }];
-}
-
-function aggregateFallOrder(msStats) {
-  var total = 0;
-  var cats = { no_fall: { count: 0, wins: 0 }, first_fall: { count: 0, wins: 0 }, second_fall: { count: 0, wins: 0 }, same_time: { count: 0, wins: 0 } };
-  Object.keys(msStats).forEach(function (ms) {
-    var fo = msStats[ms].fall_order;
-    if (!fo) return;
-    total += fo.total || 0;
-    ['no_fall', 'first_fall', 'second_fall', 'same_time'].forEach(function (k) {
-      if (fo[k]) {
-        cats[k].count += fo[k].count || 0;
-        cats[k].wins += Math.round((fo[k].win_rate || 0) / 100 * (fo[k].count || 0));
-      }
-    });
-  });
-  if (total === 0) return null;
-  function mkCat(c) {
-    return { count: c.count, rate: total > 0 ? (c.count / total * 100).toFixed(1) : '0.0', win_rate: c.count > 0 ? c.wins / c.count * 100 : 0, avg_dmg_given: null, avg_dmg_taken: null, dmg_efficiency: null };
-  }
-  return { total: total, no_fall: mkCat(cats.no_fall), first_fall: mkCat(cats.first_fall), second_fall: mkCat(cats.second_fall), same_time: mkCat(cats.same_time) };
-}
-
-function aggregateDmgContribution(msStats) {
-  var totalMatches = 0, totalContrib = 0, totalWinContrib = 0, totalWinMatches = 0, totalLoseContrib = 0, totalLoseMatches = 0;
-  Object.keys(msStats).forEach(function (ms) {
-    var dc = msStats[ms].dmg_contribution;
-    if (!dc || !dc.by_cost) return;
-    dc.by_cost.forEach(function (c) {
-      var bs = msStats[ms].basic_stats;
-      var wr = bs ? bs.win_rate / 100 : 0.5;
-      var wm = Math.round(c.matches * wr);
-      totalMatches += c.matches;
-      totalContrib += (c.avg_contribution || 0) * c.matches;
-      if (c.avg_win_contribution != null) {
-        totalWinContrib += c.avg_win_contribution * wm;
-        totalWinMatches += wm;
-      }
-      if (c.avg_lose_contribution != null) {
-        totalLoseContrib += c.avg_lose_contribution * (c.matches - wm);
-        totalLoseMatches += c.matches - wm;
-      }
-    });
-  });
-  if (totalMatches === 0) return null;
-  return { by_cost: [{ matches: totalMatches, avg_contribution: totalContrib / totalMatches, avg_win_contribution: totalWinMatches > 0 ? totalWinContrib / totalWinMatches : null, avg_lose_contribution: totalLoseMatches > 0 ? totalLoseContrib / totalLoseMatches : null }] };
-}
-
-function aggregateBurstCount(msStats) {
-  var countMap = {};
-  Object.keys(msStats).forEach(function (ms) {
-    var bc = msStats[ms].burst_count;
-    if (!bc || !bc.by_count) return;
-    bc.by_count.forEach(function (c) {
-      if (!countMap[c.label]) countMap[c.label] = { label: c.label, matches: 0, wins: 0 };
-      countMap[c.label].matches += c.matches;
-      countMap[c.label].wins += Math.round(c.win_rate / 100 * c.matches);
-    });
-  });
-  var byCount = Object.values(countMap).map(function (c) {
-    return { label: c.label, matches: c.matches, win_rate: c.matches > 0 ? c.wins / c.matches * 100 : 0 };
-  });
-  if (!byCount.length) return null;
-  return { by_count: byCount };
-}
-
-function aggregateBurstHoldDeath(msStats) {
-  var deathMap = {};
-  var noHoldWins = 0, noHoldTotal = 0, total = 0;
-  Object.keys(msStats).forEach(function (ms) {
-    var bh = msStats[ms].burst_hold_death;
-    if (!bh) return;
-    total += bh.total || 0;
-    if (bh.no_hold) {
-      noHoldTotal += bh.no_hold.count || 0;
-      noHoldWins += Math.round((bh.no_hold.win_rate || 0) / 100 * (bh.no_hold.count || 0));
-    }
-    (bh.by_death || []).forEach(function (d) {
-      if (!deathMap[d.label]) deathMap[d.label] = { label: d.label, count: 0, wins: 0 };
-      deathMap[d.label].count += d.count;
-      deathMap[d.label].wins += Math.round(d.win_rate / 100 * d.count);
-    });
-  });
-  if (total === 0) return null;
-  var byDeath = Object.values(deathMap).map(function (d) {
-    return { label: d.label, count: d.count, rate: (d.count / total * 100).toFixed(1), win_rate: d.count > 0 ? d.wins / d.count * 100 : 0 };
-  });
-  return { total: total, by_death: byDeath, no_hold: { count: noHoldTotal, rate: (noHoldTotal / total * 100).toFixed(1), win_rate: noHoldTotal > 0 ? noHoldWins / noHoldTotal * 100 : 0 } };
-}
-
-function aggregateEnemyMatchup(msStats) {
-  var enemyMap = {};
-  Object.keys(msStats).forEach(function (ms) {
-    var em = msStats[ms].enemy_matchup;
-    if (!em) return;
-    ['strong', 'weak', 'even'].forEach(function (cat) {
-      (em[cat] || []).forEach(function (e) {
-        if (!enemyMap[e.ms]) enemyMap[e.ms] = { ms: e.ms, matches: 0, wins: 0, dmgGiven: 0, dmgTaken: 0 };
-        enemyMap[e.ms].matches += e.matches;
-        enemyMap[e.ms].wins += Math.round(e.win_rate / 100 * e.matches);
-        enemyMap[e.ms].dmgGiven += (e.avg_dmg_given || 0) * e.matches;
-        enemyMap[e.ms].dmgTaken += (e.avg_dmg_taken || 0) * e.matches;
-      });
-    });
-  });
-  var all = Object.values(enemyMap).map(function (e) {
-    var wr = e.matches > 0 ? e.wins / e.matches * 100 : 0;
-    var dg = e.matches > 0 ? e.dmgGiven / e.matches : 0;
-    var dt = e.matches > 0 ? e.dmgTaken / e.matches : 0;
-    return { ms: e.ms, matches: e.matches, win_rate: wr, avg_dmg_given: dg, avg_dmg_taken: dt, dmg_efficiency: dt > 0 ? dg / dt : 0 };
-  });
-  return {
-    strong: all.filter(function (e) { return e.win_rate >= 60; }).sort(function (a, b) { return b.matches - a.matches; }),
-    weak: all.filter(function (e) { return e.win_rate <= 40; }).sort(function (a, b) { return b.matches - a.matches; }),
-    even: all.filter(function (e) { return e.win_rate > 40 && e.win_rate < 60; }).sort(function (a, b) { return b.matches - a.matches; }),
-  };
-}
-
-function aggregatePartner(msStats) {
-  var partnerMap = {};
-  Object.keys(msStats).forEach(function (ms) {
-    var partners = msStats[ms].partner;
-    if (!partners) return;
-    partners.forEach(function (p) {
-      if (!partnerMap[p.ms]) partnerMap[p.ms] = { ms: p.ms, matches: 0, wins: 0, de: 0 };
-      partnerMap[p.ms].matches += p.matches;
-      partnerMap[p.ms].wins += Math.round(p.win_rate / 100 * p.matches);
-      partnerMap[p.ms].de += (p.dmg_efficiency || 0) * p.matches;
-    });
-  });
-  return Object.values(partnerMap).map(function (p) {
-    return { ms: p.ms, matches: p.matches, win_rate: p.matches > 0 ? p.wins / p.matches * 100 : 0, dmg_efficiency: p.matches > 0 ? p.de / p.matches : 0 };
-  }).sort(function (a, b) { return b.matches - a.matches; });
 }
 
 // --- Share area ---
